+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [2.1.2] - 2026-03-09
+### Added
+- AllaganLib.GameSheets
+  - Added ENpcBase extensions and extra ItemRow properties
+
+
 ## [2.1.1] - 2026-03-09
 ### Added
 - AllaganLib.GameSheets